+++ src/commonMain/kotlin/animatedledstrip/animationutils/predefined/merge_sort_sequential.kt
@@ -48,7 +48,7 @@ val mergeSortSequential = PredefinedAnimation(
     )
 ) { leds, params, _ ->
 
-    data class SortablePixel(val finalLocation: Int, val currentLocation: Int, val color: Long)
+    data class SortablePixel(val finalLocation: Int, val currentLocation: Int, val color: Int)
 
     val colorMap = params.colors[0].colors.mapIndexed { index, it -> Pair(index, it) }.shuffled()
         .mapIndexed { index, it -> SortablePixel(it.first, index, it.second) }.toMutableList()
